chore (api-client): rename api-client
