+++ sdks/js/packages/core/react/contexts/FrontierContext.tsx
@@ -13,7 +13,7 @@ import {
   FrontierProviderProps
 } from '../../shared/types';
 
-import { V1Beta1 } from '../../apiClient/V1Beta1';
+import { V1Beta1 } from '../../api-client/V1Beta1';
 import {
   V1Beta1AuthStrategy,
   V1Beta1BillingAccount,
@@ -23,7 +23,7 @@ import {
   V1Beta1Plan,
   V1Beta1Subscription,
   V1Beta1User
-} from '../../apiClient/data-contracts';
+} from '../../api-client/data-contracts';
 import Frontier from '../frontier';
 import {
   getActiveSubscription,
