+++ sdks/js/packages/core/src/index.ts
@@ -1,5 +1,5 @@
-import { V1Beta1 } from '../apiClient/V1Beta1';
+import { V1Beta1 } from '../api-client/V1Beta1';
 
-export * from '../apiClient/V1Beta1';
-export * from '../apiClient/data-contracts';
+export * from '../api-client/V1Beta1';
+export * from '../api-client/data-contracts';
 export const FrontierClient = V1Beta1;
