+++ sdks/js/packages/sdk-demo/src/api/frontier.ts
@@ -1,4 +1,4 @@
-import { ApiClient } from '@raystack/frontier/apiClient'
+import { ApiClient } from '@raystack/frontier/api-client'
 import config from '@/config/frontier';
 
 const client = new ApiClient({
